# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rifa API Docs",
        "description": "API para gerenciamento de rifas",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "API Licence",
            "url": "https://github.com/felipesilva15/rifa-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/participant": {
            "get": {
                "tags": [
                    "Participant"
                ],
                "summary": "List all participants",
                "operationId": "613cfc5ad403d785ccd0ca182521cf53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Participant ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Participant name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Participant e-mail",
                        "schema": {
                            "type": "integer",
                            "format": "email"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "Participant phone number",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Participant list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Participant"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Participant"
                ],
                "summary": "Registers an participant",
                "operationId": "8b60cb847f123d514a6579036cd892a6",
                "requestBody": {
                    "description": "Data for creating a new participant",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Participant"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered participant data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Participant"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/participant/{id}": {
            "get": {
                "tags": [
                    "Participant"
                ],
                "summary": "List an participant by ID",
                "operationId": "f6fe5be57be1033803f6e7c9705ec793",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Participant data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Participant"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Participant"
                ],
                "summary": "Update an participant",
                "operationId": "3d30151742812d3e92ac70451a459b68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for update participant",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Participant"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated participant data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Participant"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Participant"
                ],
                "summary": "Deletes an participant",
                "operationId": "f31c019aac2e1648105a3710a933cace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Response"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error ocorried."
                    },
                    "route": {
                        "type": "string",
                        "example": "api/error"
                    }
                },
                "type": "object"
            },
            "Participant": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 80,
                        "minLength": 3,
                        "example": "Felipe Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 254,
                        "example": "felipe.allware@gmail.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "maxLength": 11,
                        "example": "11956524859"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T03:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T03:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}